+++ src/main/java/com/yooksi/betterarchery/common/CommonProxy.java
@@ -2,13 +2,31 @@ package com.yooksi.betterarchery.common;
 
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+import net.minecraftforge.fml.common.registry.GameRegistry;
 
 public class CommonProxy 
 {
 	/** Called by {@link GenericMod#preInit } in the preInit phase of mod loading. */
 	public void preInit(FMLPreInitializationEvent event) 
 	{	
-		// Read your config, create blocks, items, etc, and register them with the GameRegistry.
+		registerResources();
+	}
+	
+	/** 
+	 *  Register all items and blocks from the resource library with Forge. <br>
+	 *  This is called when the mod is passing the pre-initialization phase.
+	 */
+	private void registerResources() 
+	{
+		Logger.info("Preparing to register item and block instances...");
+		Logger.info("Finished registering object instances. ");
+	}
+	
+	private static <T extends net.minecraft.item.Item> T registerItem(T item, String name) 
+	{
+		item.setUnlocalizedName(name);
+		item.setRegistryName(name);
+		return GameRegistry.register(item);
 	}
 	
 	/** Called by {@link GenericMod#init } in the init phase of mod loading. */