+++ src/main/java/com/yooksi/betterarchery/common/ResourceLibrary.java
@@ -1,0 +1,11 @@
+package com.yooksi.betterarchery.common;
+
+
+/** This class is a storage of unique block and item instances placed here for ease of access. */
+public class ResourceLibrary
+{
+	
+	static 
+	{
+	}
+}